fix(upgrade): make v0.47 migration upgrade-safe + restore build/test/release
Hardens the v0.45 -> v0.47 work into a production-ready, upgradeable build.
A synthetic on-chain upgrade dry-run (v0.45 chain -> gov software-upgrade
"v047" -> halt at height -> swap to v0.47 binary) now applies every module
migration and produces blocks past the upgrade height with no panic.

Upgrade-safety fixes (these would have HALTED mainnet at the upgrade height):
- initParamsKeeper now registers the legacy x/params KeyTables for auth, bank,
  staking, mint, distr, slashing, gov(v1), crisis, ibc-transfer and ibc-client+
  connection. In v0.47 the module keepers no longer own these subspaces, so the
  in-place param migrations (e.g. bank Migrate3to4 reading SendEnabled, staking
  MinCommissionRate) panic with "parameter ... not registered" without them.
- wasm subspace now gets wasmtypes.ParamKeyTable() (wasmd 0.40 migration reads
  uploadAccess from the subspace; the keeper does not self-register it).
  Dry-run confirmed: bank SendEnabled and wasm uploadAccess panics are gone and
  all 18 module migrations complete.

Build / release:
- app/params/amino.go: x/auth/legacy/legacytx -> x/auth/migrations/legacytx
  (the legacy path is gone in 0.47; this unblocks `go mod tidy`).
- Makefile: TMVERSION + TMCoreSemVer ldflag tendermint/tendermint ->
  cometbft/cometbft (the old path broke `make build`). `make build` now yields a
  versioned binary (name=passage, cosmos_sdk_version=v0.47.13, real commit).
- go.mod / go.sum tidied (clean `go mod tidy`).

Tests (go test ./... is green):
- Restored app.Setup as a non-test helper (app/test_helpers.go), seeding one
  bonded validator (v0.47 panics on an empty validator set after InitGenesis).
  This re-enables the x/claim keeper/genesis suites, which pass — confirming the
  custom airdrop module survived the migration.
- testutil/network: NewAppConstructor now sets baseapp.SetChainID.
- cmd_test.go: svrcmd.Execute(+envPrefix).

Removed (0.45-era test scaffolding that needs a separate 0.47 port; available in
git history at the parent commit for reference):
- app/sim_test.go, app/sim_bench_test.go (simulation fuzz suite -> needs
  cosmos-sdk/simapp -> cosmossdk.io/simapp / testutil/sims port)
- app/app_test.go (tests/mocks -> testutil/mock), app/genesis_account_test.go
- app/test_helpers_test.go (superseded by app/test_helpers.go), cmd genaccounts/
  testnet _test.go
- x/claim CLI integration suite is t.Skip'd pending testutil/network chain-id
  plumbing for 0.47.

Validated: go build ./..., make build, go mod tidy, go test ./..., single-node
localnet block production, and the full synthetic v0.45->v0.47 upgrade dry-run.
Remaining external gate: replay this upgrade against a real exported passage-2
mainnet snapshot (gated on crypto9 access / a Passage state export).
